Refactored transfomers dependencies into the transformation context
