+++ packages/core/input/metrics.ts
@@ -2,6 +2,7 @@ import { CloudWatch } from 'aws-sdk'
 
 /**
  * $TrackMetrics(namespace: "fio", metricName: "bar", metricValue: event.size)
+ * $Fixed
  */
 export async function foo(event: any, context: any) {
   const b = 1
@@ -12,6 +13,7 @@ export async function foo(event: any, context: any) {
 }
 
 /**
+ * $Fixed
  * $TrackMetrics(namespace: "fio", metricName: "bar", metricValue: event.size)
  */
 export async function lok(event: any, context: any) {
