+++ packages/core/input/test.ts
@@ -1,30 +1,15 @@
-/** $Fixed */
-server.get('/ping', async () => {
-  return 'pong\n'
-})
-/** $Ignored
- * */
-server.listen({ port: 8080 }, (err: any, address: any) => {
-  if (err) {
-    console.error(err)
-    process.exit(1)
-  }
-  console.log(`Server listening at ${address}`)
-})
-console.log('testing')
-
-export const a = () => {}
+/**
+ * $Fixed
+ * $TrackMetrics(namespace: "fio", metricName: "bar", metricValue: event.size)
+ */
+export async function t1(event: any, context: any) {
+  return context
+}
 
 /**
- * $HttpGet
+ * $Fixed
+ * $TrackMetrics(namespace: "fio", metricName: "bar", metricValue: event.size)
  */
-export const b = () => {}
-/** $Fixed */
-export function bar() {}
-/**
- * $Ignored
- * dolor sit amet
- * Lorem ipsum
- * @returns 1
- */
-export const c = 3
+export async function t2(event: any, context: any) {
+  return event
+}
