+++ packages/core/src/index.ts
@@ -3,7 +3,7 @@ import { transpile } from './transpile'
 console.log()
 
 transpile({
-  files: ['input/metrics.ts'],
+  files: ['input/metrics.ts', 'input/test.ts'],
   serverlessConfigPath: 'input/serverless.yml',
   outputDirectory: 'output',
 })
