+++ packages/core/src/transformers/http.ts
@@ -1,29 +1,30 @@
 import ts from 'typescript'
 import { parse as parseFileName } from 'path'
-import { ServerlessConfigFunctions } from '../transpile'
-import type { AnnotationArguments } from '../annotations'
+import type { Annotation } from '../annotations'
 import { reportErrorAt } from '../report'
 
 export function httpTransfomer(
   node: ts.FunctionDeclaration,
-  functionDetails: ServerlessConfigFunctions,
-  annotationArgs: AnnotationArguments<'HttpApi'> | undefined
+  context: ts.TransformationContext,
+  annotation: Annotation<'HttpApi'>
 ): void {
   const nodeName = node.name?.getText()
   if (!nodeName) return
 
+  const annotationArgs = annotation.args
+
   if (!annotationArgs || !annotationArgs.method || !annotationArgs.path) {
     return reportErrorAt(
-      "'$HttpApi' must receive both 'method' and 'path' as parameters",
+      `$${annotation?.name}' must receive both 'method' and 'path' as parameters`,
       nodeName,
       node
     )
   }
 
-  const details = functionDetails.get(nodeName)
+  const details = context.slsFunctionDetails.get(nodeName)
 
   if (!details || !details.handler) {
-    functionDetails.set(nodeName, {
+    context.slsFunctionDetails.set(nodeName, {
       handler:
         parseFileName(node.getSourceFile().fileName).name + '.' + nodeName,
       events: [{ httpApi: annotationArgs }],
