+++ packages/core/src/transformers/index.ts
@@ -2,18 +2,45 @@ import ts from 'typescript'
 import { annotationNameEquals } from '../annotations'
 import { isNodeExported } from '../node'
 import { parseAnnotation } from '../parse'
-import type { SourceFileImports, ServerlessConfigFunctions } from '../transpile'
 import { trackMetricsTransformer } from './track-metrics'
 import { fixedTransfomer } from './fixed'
 import { httpTransfomer } from './http'
 import { scheduledTransfomer } from './scheduled'
+import type { ServerlessConfigFunctions } from '../transpile'
+
+/** This transformer maps all sub-transformers*/
+export function superTransformer(
+  checker: ts.TypeChecker,
+  functionDetails: ServerlessConfigFunctions
+): ts.TransformerFactory<ts.SourceFile> {
+  return (context) => {
+    context.slsFunctionDetails = functionDetails
+
+    return (sourceFile) => {
+      context.imports = new Set()
+
+      const visitor: ts.Visitor = (node) => {
+        if (ts.isImportDeclaration(node))
+          context.imports.add(node.moduleSpecifier.getText())
+
+        const res = mainTransfomer(node, checker, context)
+
+        // if the function transformation was successful, return the new node...
+        if (res) return res
+
+        // ...otherwise, keep traversing the AST
+        return ts.visitEachChild(node, visitor, context)
+      }
+
+      return ts.visitNode(sourceFile, visitor)
+    }
+  }
+}
 
 function mainTransfomer(
   node: ts.Node,
   checker: ts.TypeChecker,
-  context: ts.TransformationContext,
-  functionDetails: ServerlessConfigFunctions,
-  imports: SourceFileImports
+  context: ts.TransformationContext
 ): ts.Node | undefined {
   // Only consider exported function declarations nodes
   if (!isNodeExported(node) || !ts.isFunctionDeclaration(node)) return
@@ -33,50 +60,15 @@ function mainTransfomer(
     if (!parsedAnnotation) continue
 
     if (annotationNameEquals(parsedAnnotation, 'Fixed')) {
-      res = fixedTransfomer(
-        node,
-        context,
-        functionDetails,
-        parsedAnnotation.args
-      )
+      res = fixedTransfomer(node, context, parsedAnnotation)
     } else if (annotationNameEquals(parsedAnnotation, 'TrackMetrics')) {
-      res = trackMetricsTransformer(node, imports, parsedAnnotation)
+      res = trackMetricsTransformer(node, context, parsedAnnotation)
     } else if (annotationNameEquals(parsedAnnotation, 'HttpApi')) {
-      httpTransfomer(node, functionDetails, parsedAnnotation.args)
+      httpTransfomer(node, context, parsedAnnotation)
     } else if (annotationNameEquals(parsedAnnotation, 'Scheduled')) {
-      scheduledTransfomer(node, functionDetails, parsedAnnotation.args)
+      scheduledTransfomer(node, context, parsedAnnotation)
     }
   }
 
   return res
-}
-
-/** This transformer maps all sub-transformers*/
-export function superTransformer(
-  checker: ts.TypeChecker,
-  functionDetails: ServerlessConfigFunctions,
-  imports: SourceFileImports
-): ts.TransformerFactory<ts.SourceFile> {
-  return (context) => (sourceFile) => {
-    const visitor: ts.Visitor = (node) => {
-      if (ts.isImportDeclaration(node))
-        imports.set(sourceFile.fileName, node.moduleSpecifier.getText())
-
-      const res = mainTransfomer(
-        node,
-        checker,
-        context,
-        functionDetails,
-        imports
-      )
-
-      // if the function transformation was successful, return the new node...
-      if (res) return res
-
-      // ...otherwise, keep traversing the AST
-      return ts.visitEachChild(node, visitor, context)
-    }
-
-    return ts.visitNode(sourceFile, visitor)
-  }
 }
